+++ plano/src/com/hendraanggrian/plano/Size.kt
@@ -3,9 +3,9 @@ package com.hendraanggrian.plano
 /** Represents a dimension of a paper. */
 interface Size {
 
-    val width: Double
+    val width: Float
 
-    val height: Double
+    val height: Float
 
     val dimension: String get() = "${width.clean()} x ${height.clean()}"
 }
